Fix those errors with group handling
Good news is I've found the error source: I was querying tables not in the way I've been filling them.

Bad news: all that code is rather useless. Set autonomous seems to be synonymous to disband.
Maybe you'll be able to use the idea behind tha hunk of code to have a delayed vampire spawn.